put CRNN code into a shared class
